cobol: Correct diagnostic strings for 32-bit builds.
Avoid %z for printf-family.  Cast pid_t to long.  Avoid use of YYUNDEF
for old Bison versions.

	PR cobol/120621

gcc/cobol/ChangeLog:

	* genapi.cc (parser_compile_ecs): Cast argument to unsigned long.
	(parser_compile_dcls): Same.
	(parser_division): RAII.
	(inspect_tally): Cast argument to unsigned long.
	* lexio.cc (cdftext::lex_open): Cast pid_t to long.
	* parse.y: hard-code values for old versions of Bison, and message format.
	* scan_ante.h (wait_for_the_child): Cast pid_t to long.

+++ gcc/cobol/scan_ante.h
@@ -824,17 +824,20 @@ wait_for_the_child(void) {
   }
 
   if( WIFSIGNALED(status) ) {
-    yywarn( "process %d terminated by %s", pid, strsignal(WTERMSIG(status)) );
+    yywarn( "process %ld terminated by %s", 
+	    static_cast<long>(pid), strsignal(WTERMSIG(status)) );
     return false;
   }
   if( WIFEXITED(status) ) {
     if( WEXITSTATUS(status) != 0 ) {
-      yywarn("process %d exited with status %d", pid, status);
+      yywarn("process %ld exited with status %d",
+             static_cast<long>(pid), status);
       return false;
     }
   }
   if( yy_flex_debug ) {
-    yywarn("process %d exited with status %d", pid, status);
+    yywarn("process %ld exited with status %d",
+           static_cast<long>(pid), status);
   }
   return true;
 }
